perf(array): move push to CoreMongooseArray base class
Re: #7798
