+++ lib/types/documentarray.js
@@ -20,6 +20,8 @@ const arrayPathSymbol = require('../helpers/symbols').arrayPathSymbol;
 const arraySchemaSymbol = require('../helpers/symbols').arraySchemaSymbol;
 const documentArrayParent = require('../helpers/symbols').documentArrayParent;
 
+const _basePush = Array.prototype.push;
+
 /**
  * DocumentArray constructor
  *
@@ -52,7 +54,7 @@ function MongooseDocumentArray(values, path, doc) {
       arr[arrayAtomicsSymbol] = Object.assign({}, values[arrayAtomicsSymbol]);
     }
     values.forEach(v => {
-      arr.push(v);
+      _basePush.call(arr, v);
     });
   }
   arr[arrayPathSymbol] = path;
